Create a custom static topbar for total
Add a static bar at the top of the screen which contains the calculated
total as well as links to modals and to clear the screen. This replaces
the old bar which was created with the bootstrap navbar class as well as
the old area for displaying the total which was located beneath the
currency form.