Create separate template for money divs
Create a function for generating the template string of HTML that will
be turned into a div of the .money class. This allows the MakeCurrency()
function the be much cleaner and better organizes code by putting
everything that gets generated into HTML in one place.

There exist a few other minor changes to how the helper functions are
organized but none of these should affect the results of the code.