Prevent modals from shifting topbar
Event listeners have been added which modify the margin for the top bar
when modals cause the scrollbar to go away. This prevents the top bar
from erratically jumping back and forth as modals are opened/closed.

The measureScrollbar function has been coppied (with some modifications
to localize it) from the bootstrap source to ensure calculations stay
consistant.

Also made a couple of stylistic corrections of minor things that I came
across.